pyproject.toml
[project]
name = "irs_bot"
version = "0.1.0"
description = "A RAG based chatbot for the Introduction to Scientific Reasoning University of Milan course"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==1.0.6",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "durationpy==0.9",
    "fastapi==0.115.9",
    "filelock==3.18.0",
    "flask-cors>=5.0.1",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "google-auth==2.39.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kubernetes==32.0.1",
    "langchain==0.3.24",
    "langchain-community==0.3.22",
    "langchain-core==0.3.55",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.33",
    "markdown-it-py==3.0.0",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mypy-extensions==1.1.0",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "onnxruntime==1.21.1",
    "openai==1.75.0",
    "opentelemetry-api==1.32.1",
    "opentelemetry-exporter-otlp-proto-common==1.32.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.32.1",
    "opentelemetry-instrumentation==0.53b1",
    "opentelemetry-instrumentation-asgi==0.53b1",
    "opentelemetry-instrumentation-fastapi==0.53b1",
    "opentelemetry-proto==1.32.1",
    "opentelemetry-sdk==1.32.1",
    "opentelemetry-semantic-conventions==0.53b1",
    "opentelemetry-util-http==0.53b1",
    "orjson==3.10.16",
    "overrides==7.7.0",
    "packaging==24.2",
    "posthog==3.25.0",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.45.3",
    "sympy==1.13.3",
    "tenacity==9.1.2",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
